Move more types to internal
