+++ packages/config/src/internalTypes.ts
@@ -3,19 +3,26 @@ import type {
   Badge,
   ChainConfig,
   CustomDa,
-  Layer2Config,
+  Layer2FinalityConfig,
   Layer2FinalityDisplay,
+  Layer2TxConfig,
   Milestone,
+  ProjectActivityConfig,
   ProjectContracts,
   ProjectDataAvailability,
+  ProjectDaTrackingConfig,
   ProjectDiscoveryInfo,
+  ProjectEscrow,
+  ProjectLinks,
+  ProjectLivenessConfig,
   ProjectLivenessInfo,
   ProjectPermissions,
   ReasonForBeingInOther,
   ScalingProjectCapability,
-  ScalingProjectConfig,
-  ScalingProjectDisplay,
+  ScalingProjectCategory,
+  ScalingProjectPurpose,
   ScalingProjectRiskView,
+  ScalingProjectStack,
   ScalingProjectStateDerivation,
   ScalingProjectStateValidation,
   ScalingProjectTechnology,
@@ -70,6 +77,7 @@ export interface ScalingProject {
   /** Discodrive markers */
   discoveryInfo?: ProjectDiscoveryInfo
 }
+
 export interface Layer2 extends ScalingProject {
   type: 'layer2'
   display: Layer2Display
@@ -85,10 +93,68 @@ export interface Layer2Display extends ScalingProjectDisplay {
   /** Warning for Costs */
   costsWarning?: WarningWithSentiment
 }
+
 export interface Layer3 extends ScalingProject {
   type: 'layer3'
   /** ProjectId of hostChain */
   hostChain: ProjectId
   /** Stacked risk view values for this layer3 and it's base chain */
   stackedRiskView: ScalingProjectRiskView
+}
+
+export interface ScalingProjectConfig {
+  /** Associated tokens are marked on TVL breakdown -- "associated token accounts for X% of TVL" */
+  associatedTokens?: string[]
+  /** List of contracts in which L1 funds are locked */
+  escrows: ProjectEscrow[]
+  /** API parameters used to get transaction count */
+  activityConfig?: ProjectActivityConfig
+  /** Data availability tracking config */
+  daTracking?: ProjectDaTrackingConfig[]
+}
+
+export interface Layer2Config extends ScalingProjectConfig {
+  /** List of transactions that are tracked by our backend */
+  trackedTxs?: Layer2TxConfig[]
+  /** Configuration for getting liveness data */
+  liveness?: ProjectLivenessConfig
+  /** Configuration for getting finality data */
+  finality?: Layer2FinalityConfig
+}
+
+export interface ScalingProjectDisplay {
+  /** Name of the scaling project, will be used as a display name on the website */
+  name: string
+  /** Short name of the scaling project, will be used in some places on the website as a display name */
+  shortName?: string
+  /** Url friendly scaling project name, will be used in website urls */
+  slug: string
+  /** Name of the category the scaling project belongs to */
+  category: ScalingProjectCategory
+  /** Technological stack */
+  stack?: ScalingProjectStack
+  /** A warning displayed in the header of the project. Also will be displayed as yellow shield next to project name (table view) */
+  headerWarning?: string
+  /** Warning for TVL */
+  tvlWarning?: WarningWithSentiment
+  /** A warning displayed above the description of the project */
+  warning?: string
+  /** Project raw with red warning will turn into red, and there will be red warning icon with this message */
+  redWarning?: string
+  /** A few sentences describing the scaling project */
+  description: string
+  /** Detailed description of the scaling project, will be visible in detailed description section */
+  detailedDescription?: string
+  /** A short (<20 characters) description of the use case */
+  purposes: ScalingProjectPurpose[]
+  /** List of links */
+  links: ProjectLinks
+  /** Name of the architecture image to show in the contract section if present, otherwise use slug */
+  architectureImage?: string
+  /** Name of the state validation image to show in the state validation section if present, otherwise use slug */
+  stateValidationImage?: string
+  /** Name of the upgrades and governance image to show in the upgrades and governance section if present, otherwise use slug */
+  upgradesAndGovernanceImage?: string
+  /** Name of the sequencing image to show in the sequencing section if present, otherwise use slug */
+  sequencingImage?: string
 }
