+++ packages/config/src/projects/layer2s/templates/orbitStack.ts
@@ -46,7 +46,6 @@ import type {
   ProjectUpgradeableActor,
   ReasonForBeingInOther,
   ScalingProjectCapability,
-  ScalingProjectDisplay,
   ScalingProjectPurpose,
   ScalingProjectRiskView,
   ScalingProjectStateDerivation,
@@ -56,6 +55,7 @@ import type {
   StageConfig,
   TableReadyValue,
 } from '../../../types'
+import type { ScalingProjectDisplay } from '../../../internalTypes'
 import { BADGES } from '../../badges'
 import { EXPLORER_URLS } from '../../chains/explorerUrls'
 import { OPTIMISTIC_ROLLUP_STATE_UPDATES_WARNING } from '../common/liveness'
